Reduced storage needed for golay code.
The high bite of the golay_encode23 table was always int index/32 so
there’s no real need to store it, reducing the table size by half.

The same should be doable with two tables for decode, one of uint16_t
and one of uint8_t to reduce size by a quarter. This might not be
beneficial though, since it’ll mean two memory accesses, and the encode
is helped by being able to do a shift of v, but no such (easily
calculable) exists in the decode.

Currently doesn’t compile (I assume ECC was removed because it was using
too much space in RAM once other features were added), but hopefully
it’s a useful improvement saving some room for other data.